Moving configuration files.
