+++ dces-drc-integration/src/test/java/uk/gov/justice/laa/crime/dces/integration/client/external/drc/DrcApiClientTest.java
@@ -15,7 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.reactive.function.client.WebClient;
-import uk.gov.justice.laa.crime.dces.integration.configuration.client.DrcApiWebClientConfiguration;
+import uk.gov.justice.laa.crime.dces.integration.config.DrcApiWebClientConfiguration;
 import uk.gov.justice.laa.crime.dces.integration.maatapi.config.ServicesConfiguration;
 import uk.gov.justice.laa.crime.dces.integration.model.external.SendContributionFileDataToExternalRequest;
 import uk.gov.justice.laa.crime.dces.integration.model.external.SendFdcFileDataToExternalRequest;
